feat(received requests): add received requests screen
